Number Format for balance

+++ app/src/main/java/com/fnndev/pocky/data/local/models/BankAccount.kt
@@ -8,6 +8,6 @@ data class BankAccount(
     @PrimaryKey(autoGenerate = true)
     val id: Int = 0,
     val name: String,
-    val balance: Double = 0.0,
+    val balance: Long = 0L,
 )
